requirements.txt
kubernetes==34.1.0
requests==2.32.5
PyYAML==6.0.3
semver==3.0.4
PyGitHub==2.8.1
prometheus-client==0.21.1
pytest==8.3.3
pytest-mock==3.14.0
pytest-cov==6.0.0
